requirements.txt
Appium-Python-Client~=5.0.0
selenium~=4.30.0
openai~=1.70.0
numpy~=2.2.4
sounddevice~=0.5.1
soundfile~=0.13.1
webrtcvad~=2.0.10
groq~=0.22.0
python-dotenv
